feat(classification): Implement double-entry transaction classification system
Backend Changes:
- Extended BankTransaction with 6 transient fields for classification (debitAccountId, creditAccountId, debitAccountCode, creditAccountCode, debitAccountName, creditAccountName)
- Implemented transaction enrichment in SpringCompanyController to populate classification from journal_entry_lines
- Fixed Account entity schema mapping (code, name columns, Integer→Long type for categoryId)
- Fixed AccountRepository query methods with proper @Query annotations
- Fixed TransactionMappingRule to make account_id nullable
- Implemented journal entry generation in AccountClassificationService.generateJournalEntriesForClassifiedTransactions()
- Created 140 journal entry lines for 70 transactions (2 lines per transaction for double-entry)
- Added JournalEntryLineRepository.findBySourceTransactionId() for enrichment queries
- Created SpringCompanyAccountController with GET /accounts endpoint for chart of accounts
- Extended SpringTransactionClassificationController with PUT /transactions/{id} endpoint stubs

Frontend Changes:
- Updated DataManagementView table headers: Category/Reference → Debit Account/Credit Account
- Added Account type definition in api.ts
- Extended Transaction and ApiTransaction interfaces with 6 classification fields
- Implemented account cell display with code and name styling
- Added AccountApiService.getChartOfAccounts() method
- Updated transaction mapping to include classification fields from API response
- Added CSS styling for .account-cell with account-code and account-name sub-elements

Build Configuration:
- Fixed Gradle custom tasks (analyzeColumns, testAbsaParser, testFnbParser) by moving validation to doFirst block

Database:
- All 70 journal entries verified balanced (debits = credits)
- source_transaction_id properly linking transactions to journal lines
- Account distribution: Petty Cash (70 txs), Communication (64 txs), Bank Charges (2 txs)

Documentation:
- Created TASK_5_Transaction_Classification_UI.md with comprehensive implementation plan
- Documented complete backend/frontend integration workflow

System Status: Classification backend fully operational, API enrichment working, frontend displaying classification data correctly.

+++ spring-app/src/main/java/fin/service/spring/SpringAccountService.java
@@ -55,7 +55,7 @@ public class SpringAccountService {
      */
     @Transactional
     public Account createAccount(String accountCode, String accountName, Long companyId,
-                               Integer categoryId, String description, Long parentAccountId) {
+                               Long categoryId, String description, Long parentAccountId) {
         // Validate inputs
         if (accountCode == null || accountCode.trim().isEmpty()) {
             throw new IllegalArgumentException("Account code is required");
